+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [1.14.3](https://github.com/ogcio/govie-ds/compare/docs-v1.14.2...docs-v1.14.3) (2025-08-19)
+
+
+### Bug Fixes
+
+* toggle menu label visibility ([#682](https://github.com/ogcio/govie-ds/issues/682)) ([4bcf99b](https://github.com/ogcio/govie-ds/commit/4bcf99bf83bf10878500eb6be2e3833fd285961d))
+* update input text component to accept react node for icon start/end ([#683](https://github.com/ogcio/govie-ds/issues/683)) ([10e3e19](https://github.com/ogcio/govie-ds/commit/10e3e1953e4c02f827242af34c1703a8d0c7d469))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @ogcio/design-system-react bumped to 1.15.2
+    * @ogcio/design-system-tailwind bumped to 1.13.3
+
 ## [1.14.2](https://github.com/ogcio/govie-ds/compare/docs-v1.14.1...docs-v1.14.2) (2025-08-13)